Inline list.js (again) and search hotfix
I thought it makes HTML document too big to load and makes page load slower, but I think I was wrong about it because of the total blocking time aspect.

And I was a total idiot by removing that .name value completely, because that was for List.js. Renamed to as "n".